backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pymysql==1.1.1
python-dotenv==1.0.1
